Move Money to its own file
Money is not just used as part of a ShippingOption. It’s used as its own type in its own right all the time. Let’s move it.
